refactor: split HttpServer.scala into modular route files
Extract DTOs, Circe codecs, and routes into separate files:
- dto/Dto.scala: ValidationErrorDto hierarchy, auth and army DTOs
- CirceCodecs.scala: UUID, WargearSelection, ArmyUnit, Army codecs
- routes/AuthRoutes.scala: auth and invite endpoints
- routes/FactionRoutes.scala: faction data endpoints
- routes/ArmyRoutes.scala: army CRUD and validation
- routes/DatasheetRoutes.scala: datasheet detail endpoints

HttpServer.scala now handles only server setup and route composition.
